+++ uw-cache/src/main/java/uw/cache/GlobalCache.java
@@ -1,14 +1,18 @@
 package uw.cache;
 
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.data.redis.core.Cursor;
 import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.core.ScanOptions;
 import org.springframework.data.redis.core.ValueOperations;
 import uw.cache.util.KryoUtils;
 import uw.cache.util.RedisKeyUtils;
 import uw.cache.vo.CacheProtectedValue;
 
 import java.lang.reflect.*;
+import java.util.HashSet;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
 
@@ -52,7 +56,7 @@ public class GlobalCache {
      */
     private static final String REDIS_PREFIX = "uw-cache:";
 
-    private static final Logger logger = LoggerFactory.getLogger( GlobalCache.class );
+    private static final Logger logger = LoggerFactory.getLogger(GlobalCache.class);
 
     /**
      * 这里使用ValueOperations<String, String>是有原因的 虽然<K, V>可以泛型 但是泛型也存在问题 使用时必须实例化 也就是说每个使用的地方得ValueOperations<类, 类>实例化RedisTemplate
@@ -75,8 +79,8 @@ public class GlobalCache {
      * @param value        数据
      * @param expireMillis 有效期毫秒数。
      */
-    public static <K, V> void put(Class entityClass, K key, V value, long expireMillis) {
-        put( entityClass.getSimpleName(), key, value, expireMillis );
+    public static <K, V> void put(Class<?> entityClass, K key, V value, long expireMillis) {
+        put(entityClass.getSimpleName(), key, value, expireMillis);
     }
 
     /**
@@ -86,16 +90,16 @@ public class GlobalCache {
      * @param key          主键
      * @param value        数据
      * @param expireMillis 有效期毫秒数。
-     * @param <V>
-     * @return
+     * @param <V> 数据类型
+     * @return void
      */
     public static <K, V> void put(String cacheName, K key, V value, long expireMillis) {
-        String redisKey = RedisKeyUtils.buildTypeId( REDIS_PREFIX, cacheName, key );
-        byte[] redisData = KryoUtils.serialize( value );
+        String redisKey = RedisKeyUtils.buildTypeId(REDIS_PREFIX, cacheName, key);
+        byte[] redisData = KryoUtils.serialize(value);
         if (expireMillis == 0) {
-            opsForValue.set( redisKey, redisData );
+            opsForValue.set(redisKey, redisData);
         } else {
-            opsForValue.set( redisKey, redisData, expireMillis, TimeUnit.MILLISECONDS );
+            opsForValue.set(redisKey, redisData, expireMillis, TimeUnit.MILLISECONDS);
         }
     }
 
@@ -107,11 +111,11 @@ public class GlobalCache {
      * @param key             主键
      * @param cacheDataLoader 加载数据的函数
      * @param expireMillis    有效期毫秒数。
-     * @param <V>
+     * @param <V> 数据类型
      * @return
      */
-    public static <K, V> V get(Class entityClass, K key, CacheDataLoader<K, V> cacheDataLoader, long expireMillis) {
-        return get( entityClass.getSimpleName(), key, cacheDataLoader, expireMillis );
+    public static <K, V> V get(Class<?> entityClass, K key, CacheDataLoader<K, V> cacheDataLoader, long expireMillis) {
+        return get(entityClass.getSimpleName(), key, cacheDataLoader, expireMillis);
     }
 
     /**
@@ -122,12 +126,11 @@ public class GlobalCache {
      * @param key             主键
      * @param cacheDataLoader 加载数据的函数
      * @param expireMillis    有效期毫秒数。
-     * @param <V>
+     * @param <V> 数据类型
      * @return
      */
     public static <K, V> V get(String cacheName, K key, CacheDataLoader<K, V> cacheDataLoader, long expireMillis) {
-        return get( cacheName, key, cacheDataLoader, expireMillis, DEFAULT_NULL_PROTECT_MILLIS, DEFAULT_FAIL_PROTECT_MILLIS, DEFAULT_RELOAD_INTERVAL_MILLIS,
-                DEFAULT_RELOAD_MAX_TIMES );
+        return get(cacheName, key, cacheDataLoader, expireMillis, DEFAULT_NULL_PROTECT_MILLIS, DEFAULT_FAIL_PROTECT_MILLIS, DEFAULT_RELOAD_INTERVAL_MILLIS, DEFAULT_RELOAD_MAX_TIMES);
     }
 
     /**
@@ -143,9 +146,8 @@ public class GlobalCache {
      * @param reloadMaxTimes       重载次数
      * @return
      */
-    public static <K, V> V get(Class entityClass, K key, CacheDataLoader<K, V> cacheDataLoader, long expireMillis, long nullProtectMillis, long failProtectMillis,
-                               long reloadIntervalMillis, int reloadMaxTimes) {
-        return get( entityClass.getSimpleName(), key, cacheDataLoader, expireMillis, nullProtectMillis, failProtectMillis, reloadIntervalMillis, reloadMaxTimes );
+    public static <K, V> V get(Class<?> entityClass, K key, CacheDataLoader<K, V> cacheDataLoader, long expireMillis, long nullProtectMillis, long failProtectMillis, long reloadIntervalMillis, int reloadMaxTimes) {
+        return get(entityClass.getSimpleName(), key, cacheDataLoader, expireMillis, nullProtectMillis, failProtectMillis, reloadIntervalMillis, reloadMaxTimes);
     }
 
     /**
@@ -164,9 +166,8 @@ public class GlobalCache {
      * @param reloadMaxTimes       重载次数
      * @return
      */
-    public static <K, V> V get(String cacheName, K key, CacheDataLoader<K, V> cacheDataLoader, long expireMillis, long nullProtectMillis, long failProtectMillis,
-                               long reloadIntervalMillis, int reloadMaxTimes) {
-        V value = loadWithProtectedValue( cacheName, key, cacheDataLoader, expireMillis, nullProtectMillis, failProtectMillis, reloadIntervalMillis, reloadMaxTimes );
+    public static <K, V> V get(String cacheName, K key, CacheDataLoader<K, V> cacheDataLoader, long expireMillis, long nullProtectMillis, long failProtectMillis, long reloadIntervalMillis, int reloadMaxTimes) {
+        V value = loadWithProtectedValue(cacheName, key, cacheDataLoader, expireMillis, nullProtectMillis, failProtectMillis, reloadIntervalMillis, reloadMaxTimes);
         if (value instanceof CacheProtectedValue protectValue) {
             return null;
         }
@@ -189,16 +190,15 @@ public class GlobalCache {
      * @param reloadMaxTimes       重载次数
      * @return
      */
-    public static <K, V> V loadWithProtectedValue(String cacheName, K key, CacheDataLoader<K, V> cacheDataLoader, long expireMillis, long nullProtectMillis,
-                                                  long failProtectMillis, long reloadIntervalMillis, int reloadMaxTimes) {
+    public static <K, V> V loadWithProtectedValue(String cacheName, K key, CacheDataLoader<K, V> cacheDataLoader, long expireMillis, long nullProtectMillis, long failProtectMillis, long reloadIntervalMillis, int reloadMaxTimes) {
         //组成真正的RedisKey
-        String redisKey = RedisKeyUtils.buildTypeId( REDIS_PREFIX, cacheName, key );
+        String redisKey = RedisKeyUtils.buildTypeId(REDIS_PREFIX, cacheName, key);
         // 从redis中获取value.
         byte[] redisData = null;
         try {
-            redisData = opsForValue.get( redisKey );
+            redisData = opsForValue.get(redisKey);
         } catch (Exception e) {
-            logger.error( e.getMessage(), e );
+            logger.error(e.getMessage(), e);
         }
         //对象数值。
         V value = null;
@@ -209,11 +209,11 @@ public class GlobalCache {
             } else {
                 try {
                     //正常数据
-                    value = (V) KryoUtils.deserialize( redisData, type2Class( cacheDataLoader.getValueType() ) );
+                    value = (V) KryoUtils.deserialize(redisData, type2Class(cacheDataLoader.getValueType()));
                     return value;
                 } catch (Throwable e) {
                     //正常数据
-                    value = (V) KryoUtils.deserialize( redisData, CacheProtectedValue.class );
+                    value = (V) KryoUtils.deserialize(redisData, CacheProtectedValue.class);
                     return value;
                 }
             }
@@ -222,9 +222,9 @@ public class GlobalCache {
         synchronized (redisKey.intern()) {
             // 同一jvm中执行这个会被其他线程加锁阻塞 等待那个线程释放锁后 此线程进来 尝试再去get一下值 apply方法执行正常这里就会可以get到
             try {
-                redisData = opsForValue.get( redisKey );
+                redisData = opsForValue.get(redisKey);
             } catch (Exception e) {
-                logger.error( e.getMessage(), e );
+                logger.error(e.getMessage(), e);
             }
 
             if (redisData != null) {
@@ -235,11 +235,11 @@ public class GlobalCache {
                     //正常数据
                     try {
                         //正常数据
-                        value = (V) KryoUtils.deserialize( redisData, type2Class( cacheDataLoader.getValueType() ) );
+                        value = (V) KryoUtils.deserialize(redisData, type2Class(cacheDataLoader.getValueType()));
                         return value;
                     } catch (Throwable e) {
                         //正常数据
-                        value = (V) KryoUtils.deserialize( redisData, CacheProtectedValue.class );
+                        value = (V) KryoUtils.deserialize(redisData, CacheProtectedValue.class);
                         return value;
                     }
                 }
@@ -247,36 +247,36 @@ public class GlobalCache {
             // 假如还是没有get到值 则就是apply方法执行报错了 可以尝试继续执行 再放入redis
             for (int retryTimes = 0; retryTimes < reloadMaxTimes; retryTimes++) {
                 try {
-                    value = cacheDataLoader.load( key );
+                    value = cacheDataLoader.load(key);
                     if (value == null) {
-                        value = (V) new CacheProtectedValue( nullProtectMillis );
+                        value = (V) new CacheProtectedValue(nullProtectMillis);
                         expireMillis = nullProtectMillis;
                     }
                     //正常执行就退出吧。
                     break;
                 } catch (Throwable e) {
-                    logger.error( "Global数据加载失败! cacheName:{}, key:{}, retryTimes:{}, msg:{}", cacheName, key, retryTimes, e.getMessage(), e );
+                    logger.error("Global数据加载失败! cacheName:{}, key:{}, retryTimes:{}, msg:{}", cacheName, key, retryTimes, e.getMessage(), e);
                 }
                 try {
-                    Thread.sleep( reloadIntervalMillis );
+                    Thread.sleep(reloadIntervalMillis);
                 } catch (InterruptedException ignored) {
                 }
             }
             // 如果此时还没有得到数值，说明加载彻底失败。
             if (value == null) {
-                value = (V) new CacheProtectedValue( failProtectMillis );
+                value = (V) new CacheProtectedValue(failProtectMillis);
                 expireMillis = failProtectMillis;
             }
             //序列化写库。
-            redisData = KryoUtils.serialize( value );
+            redisData = KryoUtils.serialize(value);
             try {
                 if (expireMillis == 0) {
-                    opsForValue.set( redisKey, redisData );
+                    opsForValue.set(redisKey, redisData);
                 } else {
-                    opsForValue.set( redisKey, redisData, expireMillis, TimeUnit.MILLISECONDS );
+                    opsForValue.set(redisKey, redisData, expireMillis, TimeUnit.MILLISECONDS);
                 }
             } catch (Exception e) {
-                logger.error( e.getMessage(), e );
+                logger.error(e.getMessage(), e);
             }
         }
         return value;
@@ -286,33 +286,52 @@ public class GlobalCache {
      * 删除缓存中的数据。
      *
      * @param entityClass 缓存对象类(主要用于构造cacheName)
-     * @param key
+     * @param key 缓存主键，null则全部清除
      * @return
      */
-    public static boolean invalidate(Class entityClass, Object key) {
-        return invalidate( entityClass.getSimpleName(), key );
+    public static boolean invalidate(Class<?> entityClass, Object key) {
+        return invalidate(entityClass.getSimpleName(), key);
     }
 
     /**
      * 删除缓存中的数据。
      *
-     * @param cacheName
-     * @param key
+     * @param cacheName 缓存名
+     * @param key 缓存主键，null则全部清除
      * @return
      */
     public static boolean invalidate(String cacheName, Object key) {
         //如果key是null，则清除全部。
         if (key == null) {
-            key = "*";
-            //组成真正的RedisKey
-            String redisKey = RedisKeyUtils.buildTypeId( REDIS_PREFIX, cacheName, key );
-            Set<String> keys = cacheRedisTemplate.keys( redisKey );
-            cacheRedisTemplate.delete( keys );
+            cacheRedisTemplate.delete(keys(cacheName,null));
             return true;
         } else {
-            String redisKey = RedisKeyUtils.buildTypeId( REDIS_PREFIX, cacheName, key );
-            return Boolean.TRUE.equals( cacheRedisTemplate.delete( redisKey ) );
+            String redisKey = RedisKeyUtils.buildTypeId(REDIS_PREFIX, cacheName, key);
+            return Boolean.TRUE.equals(cacheRedisTemplate.delete(redisKey));
         }
+    }
+
+    /**
+     * 删除缓存中指定前缀的Key。
+     *
+     * @param entityClass 缓存对象类(主要用于构造cacheName)
+     * @param keyPrefix key前缀，请注意key最后不用加"*"
+     * @return
+     */
+    public static boolean invalidateKeys(Class<?> entityClass, String keyPrefix) {
+        return invalidate(entityClass.getSimpleName(), keyPrefix);
+    }
+
+    /**
+     * 删除缓存中指定前缀的Key。
+     *
+     * @param cacheName 缓存名
+     * @param keyPrefix key前缀，请注意key最后不用加"*",全部清除用*即可。
+     * @return
+     */
+    public static boolean invalidatePrefix(String cacheName, String keyPrefix) {
+        cacheRedisTemplate.delete(keys(cacheName,keyPrefix));
+        return true;
     }
 
     /**
@@ -327,8 +346,8 @@ public class GlobalCache {
      * @param <V>
      * @return
      */
-    public static <K, V> V get(Class entityClass, K key, CacheDataLoader<K, V> cacheDataLoader, long expireMillis, long failProtectMillis) {
-        return get( entityClass.getSimpleName(), key, cacheDataLoader, expireMillis, failProtectMillis );
+    public static <K, V> V get(Class<?> entityClass, K key, CacheDataLoader<K, V> cacheDataLoader, long expireMillis, long failProtectMillis) {
+        return get(entityClass.getSimpleName(), key, cacheDataLoader, expireMillis, failProtectMillis);
     }
 
     /**
@@ -340,11 +359,11 @@ public class GlobalCache {
      * @param cacheDataLoader   加载数据函数
      * @param expireMillis      有效期毫秒数。
      * @param failProtectMillis 失败保护毫秒数
-     * @param <V>
+     * @param <V> 数据类型
      * @return
      */
     public static <K, V> V get(String cacheName, K key, CacheDataLoader<K, V> cacheDataLoader, long expireMillis, long failProtectMillis) {
-        return get( cacheName, key, cacheDataLoader, expireMillis, failProtectMillis, DEFAULT_NULL_PROTECT_MILLIS, DEFAULT_RELOAD_INTERVAL_MILLIS, DEFAULT_RELOAD_MAX_TIMES );
+        return get(cacheName, key, cacheDataLoader, expireMillis, failProtectMillis, DEFAULT_NULL_PROTECT_MILLIS, DEFAULT_RELOAD_INTERVAL_MILLIS, DEFAULT_RELOAD_MAX_TIMES);
     }
 
     /**
@@ -354,7 +373,7 @@ public class GlobalCache {
      * @param message 消息。
      */
     public static Long notifyMsg(String channel, Object message) {
-        return cacheRedisTemplate.convertAndSend( channel, KryoUtils.serialize( message ) );
+        return cacheRedisTemplate.convertAndSend(channel, KryoUtils.serialize(message));
     }
 
     /**
@@ -369,20 +388,54 @@ public class GlobalCache {
         } else if (type instanceof GenericArrayType) {
             // having to create an array instance to get the class is kinda nasty
             // but apparently this is a current limitation of java-reflection concerning array classes.
-            return Array.newInstance( type2Class( ((GenericArrayType) type).getGenericComponentType() ), 0 ).getClass(); // E.g. T[] -> T -> Object.class if <T> or Number.class
+            return Array.newInstance(type2Class(((GenericArrayType) type).getGenericComponentType()), 0).getClass(); // E.g. T[] -> T -> Object.class if <T> or Number.class
             // if <T extends Number & Comparable>
         } else if (type instanceof ParameterizedType) {
-            return type2Class( ((ParameterizedType) type).getRawType() ); // Eg. List<T> would return List.class
+            return type2Class(((ParameterizedType) type).getRawType()); // Eg. List<T> would return List.class
         } else if (type instanceof TypeVariable) {
             Type[] bounds = ((TypeVariable<?>) type).getBounds();
-            return bounds.length == 0 ? Object.class : type2Class( bounds[0] ); // erasure is to the left-most bound.
+            return bounds.length == 0 ? Object.class : type2Class(bounds[0]); // erasure is to the left-most bound.
         } else if (type instanceof WildcardType) {
             Type[] bounds = ((WildcardType) type).getUpperBounds();
-            return bounds.length == 0 ? Object.class : type2Class( bounds[0] ); // erasure is to the left-most upper bound.
+            return bounds.length == 0 ? Object.class : type2Class(bounds[0]); // erasure is to the left-most upper bound.
         } else {
             // throw new UnsupportedOperationException( "cannot handle type class: " + type.getClass() );
             return Object.class;
         }
+    }
+
+    /**
+     * 获取缓存中的所有key。
+     * @param entityClass 缓存对象类(主要用于构造cacheName)
+     * @param keyPrefix key前缀，请注意key最后不用加"*"
+     * @return key集合
+     */
+    public static Set<String> keys(Class<?> entityClass, String keyPrefix) {
+        return keys(entityClass.getSimpleName(), keyPrefix);
+    }
+
+    /**
+     * 获取缓存中的所有key。
+     *
+     * @param cacheName 缓存名
+     * @param keyPrefix key前缀，请注意key最后不用加"*",全部清除用*即可。
+     * @return key集合
+     */
+    public static Set<String> keys(String cacheName, String keyPrefix) {
+        if (StringUtils.isBlank(keyPrefix)) {
+            keyPrefix = "*";
+        } else {
+            keyPrefix = keyPrefix + "*";
+        }
+        int redisPrefixLength = REDIS_PREFIX.length() + cacheName.length() + 1;
+        String pattern = RedisKeyUtils.buildTypeId(REDIS_PREFIX, cacheName, keyPrefix);
+        Set<String> keys = new HashSet<>();
+        try (Cursor<String> cursor = cacheRedisTemplate.scan(ScanOptions.scanOptions().match(pattern).count(1000).build())) {
+            cursor.forEachRemaining(key -> {
+                keys.add(key.substring(redisPrefixLength));
+            });
+        }
+        return keys;
     }
 
 }