+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2025.0501.1029版本升级说明
+1. uw-cache: GlobalCache,GlobalCounter,GlobalHashSet,GlobalLocker,GlobalSortedSet 新增通过key前缀获取key列表功能。
+2. uw-cache: 优化FusionCounter相关功能，修复了数据库回写逻辑。
+3. uw-cache: 优化FusionCache相关功能，增加了keys相关方法。
+
 # 2025.0501.1028版本升级说明
 1. uw-common: DateUtils 新增stringToDate()新增自动高效解析时间戳格式。
 2. uw-log-es: 移除重复的@PreDestory指令。